refactor: extract claude-mem import from cli
Move the claude-mem import implementation into opencode_mem/commands/import_from_claude_mem.py and keep cli.py as a thin wrapper, reducing the CLI god-file size.
